requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-WTF>=1.2
email-validator>=2.1
Flask-Migrate>=4.0
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
PyMySQL>=1.1
Werkzeug>=3.0
